WIP save point for merch store and DB update and sending data to view
