+++ cptc-CPW219-eCommerceSite/Views/Home/CreateNewMerchProductDialog.cshtml
@@ -6,20 +6,16 @@
 
 
 <div class="modal" tabindex="-1" id="create-modal">
-    <div class="modal-dialog">
-        <div class="modal-content">
+    <div class="modal-dialog modal-lg">
+        <form class="modal-content" asp-action="CreateNewMerchProductDialog">
             <div class="modal-header">
-                <h5 class="modal-title">Modal title</h5>
+                <h5 class="modal-title">Create Product for Merch Store</h5>
                 <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
             </div>
             <div class="modal-body">
-                <h1>CreateNewMerchProductDialog</h1>
-
-                <h4>ProductViewModel</h4>
-                <hr />
                 <div class="row">
-                    <div class="col-md-4">
-                        <form asp-action="CreateNewMerchProductDialog">
+                    <div class="col">
+                        <div >
                             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
                             <div class="form-group">
                                 <label asp-for="Name" class="control-label"></label>
@@ -47,17 +43,19 @@
                                 <span asp-validation-for="CategoryName" class="text-danger"></span>
                             </div>
                             <div class="form-group">
-                                <input type="submit" value="Create" class="btn btn-primary" />
+                                <label asp-for="ImageFile" class="control-label"></label>
+                                <input asp-for="ImageFile" type="file" class="form-control" />
+                                <span asp-validation-for="ImageFile" class="text-danger"></span>
                             </div>
-                        </form>
+                        </div>
                     </div>
                 </div>
             </div>
             <div class="modal-footer">
                 <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-                <button type="button" class="btn btn-primary">Save changes</button>
+                <input type="submit" value="Create" class="btn btn-primary" />
             </div>
-        </div>
+        </form>
     </div>
 </div>
