+++ cptc-CPW219-eCommerceSite/Views/Home/_partials/_merch-store.cshtml
@@ -26,7 +26,7 @@
 
 
         <h2 class="text-center">Merch Store</h2>
-         <div>total: @Model.Products.Count()</div> 
+         <div>total: @Model.ProductsVM.Count()</div> 
 
         <div>
